Planning document, SQL Defaults, Populating DB
I crafted a planning document to organize my thoughts regarding the
design of the spider. I changed python3.6 to simply python3 so that
this would run on raspbian, which ships with python3.5. I modified the
SQL so that various table fields have default values, then added two
seed urls to the database. The script will no longer accept
user-provided seed urls, instead opting to scan entirely out of the
urls provided in the database. If anyone wants to add new seed urls,
they can simply implement them in the same manner as the two I've
already added. Removed the url requirement from the crawl function,
since it'll be choosing its next url either from the database or from a
queue.