Nearly there!
I finished the majority of the program, tested it, and it successfully
connected and began spidering the seed urls. In the test, it discovered
over 6,000 new domains and began scanning those, adding links along the
way. The only problem I encountered were occasional malformed links
that look like: http:///onion/s4p52h3uq4ls3ka2.onion – So I need to
sort that bug out before this will be complete. It shouldn't be
terribly hard to defeat, but I need to know what is causing that
malformation. Anyway, solid progress so far!