First Draft Complete, Ready for Testing
This is the first draft of v2. At this point it's a complete rewrite of
the original, but it employs no new concepts, with the exception of
letting the Scribe handle all writing requests.

Now, I just need to do some QA testing to ensure that this thing runs
the way it's expected to, and to iron out all the bugs before I move on
to the next phase of the operation.