add_url, set_fault, status codes
I moved the code to add a new url to an add_url function, allowing me
to add a new url to the database from anywhere in the crawl function. I
updated some places that were manually setting faults to use the
set_fault function. I updated the status codes in the 300-series to get
the redirected location and add it as a new url to the database, so
that we make sure to have the new urls available to us.