form data, fixed db-con bug
Added the table forms to the database. It stores the form fields and
examples of the various queries that are passed to those fields. Also,
I fixed a bug in the db-con script that was causing some issues with
conversion. This bug only became apparent when I created the forms
database and tried to populate it.