Streamline scans.
db_update created to update the database to reflect the new
requirements. TorSpider updated to change the way offline pages are
handled, so that they're not scanned so frequently. The longer a page
is offline, the less often it gets scanned. These changes have not yet
been tested in the wild.