Forms processed and added to DB. Bug fix, too.
The spiders now parse and process the form data they find and add it to the database. Also, I discovered there was a significant bug where I was linking form data to the wrong pages in the database. I fixed this bug, but it means all prior databases are invalid and should be wiped. This code is now ready to be bugtested prior to pull request.