Integrate form scraping. (#8)
* First steps in parsing forms.

Spiders now parse forms. As of now, the only thing they do is add the action url to the database and print the form information to the console.

* Form data ready to be processed.

All discovered form data is now ready to be processed into SQL statements and added to the database.

* Forms processed and added to DB. Bug fix, too.

The spiders now parse and process the form data they find and add it to the database. Also, I discovered there was a significant bug where I was linking form data to the wrong pages in the database. I fixed this bug, but it means all prior databases are invalid and should be wiped. This code is now ready to be bugtested prior to pull request.

* Debug.

There's a persistent Index Error when trying to get the link_domain_id on line 368. Sometimes it works, other times it doesn't. I'm trying to nail this down.

* Bug Squashed.

The problem arose when we encountered forms with actions referring to non-onion domains. We ignore those by default.

* Fixing "no select_options" bug.

System would throw an exception when it tried to assign a value to select_options before select_options exists. Fixed this bug.